requirements.main.txt
orjson>=3.7.5
pyyaml>=6.0
